Created a "personnal branch" to work on different computer without polluting master

+++ Core/src/string/string/StringUtils.cpp
@@ -2,7 +2,6 @@
 
 #include <cstring>
 #include <cstdarg>
-#include <cstdio>
 #include <memory>
 #include <sstream>
 
@@ -13,7 +12,7 @@ namespace Ra
         std::string getFileExt(const std::string& str)
         {
             std::string res;
-            ulong pos = str.find_last_of('.');
+            unsigned pos = str.find_last_of('.');
             res = pos < str.size() ? str.substr(pos) : "";
             return res;
         }
@@ -22,7 +21,7 @@ namespace Ra
         std::string getDirName(const std::string& path)
         {
             // We remove any trailing slashes.
-            ulong pos = path.find_last_not_of('/');
+            unsigned int pos = path.find_last_not_of('/');
 
             // Don't strip the last / from "/"
             if (pos == std::string::npos)
@@ -58,7 +57,7 @@ namespace Ra
         {
             std::string res;
             // We remove any trailing slashes.
-            ulong pos = path.find_last_not_of('/');
+            unsigned int pos = path.find_last_not_of('/');
             // Don't strip the last / from "/"
             if (pos == std::string::npos)
             {
@@ -82,7 +81,7 @@ namespace Ra
         int stringPrintf(std::string& str, const char* fmt, ...)
         {
             // Random guessing value from the size of the format string.
-            size_t size = strlen(fmt) * 2;
+            int size = strlen(fmt) * 2;
             int finalSize = 0;
             str.clear();
 
@@ -103,7 +102,7 @@ namespace Ra
                 // gives us the required buffer size.
                 if (finalSize >= size)
                 {
-                    size = std::max(size + 1, size_t(finalSize));
+                    size = std::max(size + 1, finalSize);
                 }
                 else
                 {
